Adding a new sandboxed application for the distortion engine.
This app will, temporarily, house the functionality of the distortion
engine. The rationale for creating a new app is this - testing
distorting sounds from the main application requires other components
tobe ready and integrated which will take a while.

This app will have a minimalistic UI for setting some parameters and
playing the distorting sounds. The code will be written in a way that
will enable (hopefully) easy merging with the main application later on.