Distortion sound files and sound playing class.
The three mp3 files added in the raw folder are (for now) the distortion
sounds that can be played. All of them are simple static sounds.

DeProxy will get the name of the sound file (without the extension) from
the intent and pass the same to the player package, where a mapping from
the name to the resource's ID defined.

I've added Javadoc for all newly added methods and comments where
necessary. There are a bunch of TODO's as well.