Added UI and parameterized playing of distortion sound(s).
UI: A group of radio buttons to select the sound to play, a text field
to enter the amount of time to play the sound (in seconds) and a button
(duh!).

Clicking the button causes a broadcast to be send, whose intent has the
name of the sound to be played and the play length. The receiver
receives the broadcast, unpacks the intent and plays the sound.

Looping is set to on for the MediaPlayer object and a timer is set to
have the sound stop after the user-entered play length.

All of the above, although happening in a different app, mimics what
must happen in the main app - this app's UI does the job of the
component of the main app responsible for initiating the distortion.
This should make it easier to integrate the DE later on.